packages/librarian/pyproject.toml
[project]
name = "librarian"
version = "0.1.0"
description = "A simple API for managing library users, books, and borrowing."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.4.0",
    "gunicorn>=23.0.0",
    "whitenoise>=6.7.0",
    "drf-yasg>=1.21.7",
    "django>=5.1.1",
    "pagekeeper",
    "bookcourier",
    "shortuuid>=1.0.13",
    "environs[django]>=11.0.0",
    "psycopg2-binary>=2.9.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
pagekeeper = { workspace = true }
bookcourier = { workspace = true }
